Add files via upload
The project is still in very early stages. Still learning a lot of stuff. These are the first files I'm working on. 

bsoupLearning is a file for the tests and my learning process using the Bsoup library.

WikiPrinter.py is the python script that will (eventually and hopefully) scrape text from a Wikipedia article.